improve(mnDiagram3_80245BA4): drop redundant u32 cast on icon-id sentinel — 94.48 -> 94.91
Compare 'r17 == 0xFFFF' instead of '(u32) r17 == 0xFFFF'. r17 is the int copy
of a u16 (0..65535) so the cast is redundant and the test is behaviorally
identical, but without it MWCC matches the target's compare/keep sequence
(addis r0,r3,0; cmplwi r0,65535; addi r17,r3,0) instead of folding differently.

+++ src/melee/mn/mndiagram3.c
@@ -227,7 +227,7 @@ void mnDiagram3_80245BA4(HSD_GObj* gobj)
                 {
                     u16 icon_id = *stat_table;
                     int r17 = icon_id;
-                    if ((u32) r17 == 0xFFFF) {
+                    if (r17 == 0xFFFF) {
                         goto next;
                     }
 
@@ -247,7 +247,8 @@ void mnDiagram3_80245BA4(HSD_GObj* gobj)
                             if (mnDiagram_IsDistanceOverflow(
                                     mnDiagram2_GetStatValue(data->is_name_mode,
                                                             stat_type,
-                                                            entity))) {
+                                                            entity)))
+                            {
                                 HSD_SisLib_803A6368(icon_text, 0x7F);
                                 goto next;
                             }
